requirements.txt
openai
pillow
tkinter
Python-IO